chart.js, docker configuration, refactoringcode

+++ src/main/java/com/grownited/controller/PaymentController.java
@@ -1,5 +1,7 @@
 package com.grownited.controller;
 
+import java.math.BigDecimal;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -23,7 +25,7 @@ public class PaymentController {
 
 	@PostMapping("charge")
 	public String chargeCreditCard(String cardNumber, String expMonth, String expYear, 
-			Double amount, @SessionAttribute("user") UserEntity user) {
+			BigDecimal amount, @SessionAttribute("user") UserEntity user) {
 
 		String email = user.getEmail();
 		paymentService.chargeCreditCard(email, cardNumber, expMonth + expYear, amount);
